Start work on builtin classes (not working yet)
