+++ src/lib.rs
@@ -43,7 +43,10 @@ impl PluginGroup for TrenchBroomPlugins {
 			.add(CorePlugin(self.0))
 			.add(fgd::FgdPlugin)
 			.add(class::QuakeClassPlugin)
-			.add(class::builtin::BuiltinClassesPlugin)
+			.add(class::builtin::BasicClassesPlugin)
+			.add(class::builtin::LightingClassesPlugin(default()))
+			.add(class::builtin::BspClassesPlugin)
+			.add(class::builtin::SolidClassesPlugin)
 			.add(config::ConfigPlugin)
 			.add(qmap::QuakeMapPlugin)
 			.add(geometry::GeometryPlugin)
